removed the center of mass stuff. by default the contigs are not sorted now or left in the order of the r2c file. introduced for the ordering a sortkey variable in DNASequence
there are still some issues: The contigs / targets are stored twice. once in a hashmap which is quite useful for adding new alignment positions, and additionally in a vector which is used to change the contig order manually.
this should probably be fixed but this is quite tricky since each representation is uses in its context..